+++ src/labs/inlab4/Parser.h
@@ -162,20 +162,20 @@ class Parser : public CommandLine<InputMatrices> {
             printLine();
             std::cout << "Lower Triangular Matrix (L):\n";
             printLine();
-            std::cout << std::setprecision(precision) << lower;
+            std::cout << std::setprecision(static_cast<int>(precision)) << lower;
             printLine();
             std::cout << "Upper Triangular Matrix (U):\n";
             printLine();
-            std::cout << std::setprecision(precision) << upper;
+            std::cout << std::setprecision(static_cast<int>(precision)) << upper;
             printLine();
             std::cout << "Constants Vector (b):\n";
             printLine();
-            std::cout << std::setprecision(precision) << inputs.constants;
+            std::cout << std::setprecision(static_cast<int>(precision)) << inputs.constants;
             printLine();
             printLine();
             std::cout << "LU Matrix:\n";
             printLine();
-            std::cout << std::setprecision(precision) << inputs.LU;
+            std::cout << std::setprecision(static_cast<int>(precision)) << inputs.LU;
         }
     }
 };
