+++ src/labs/inlab4/inlab4.h
@@ -108,11 +108,11 @@ class InLab4 : public Project<InputMatrices, Parser, OutputVector> {
             Parser::printLine();
             std::cout << "Intermediate vector y = inv(L) * b:\n";
             Parser::printLine();
-            std::cout << std::setprecision(precision) << y;
+            std::cout << std::setprecision(static_cast<int>(precision)) << y;
             Parser::printLine();
             std::cout << "Solution vector (x):\n";
             Parser::printLine();
-            std::cout << std::setprecision(precision) << x;
+            std::cout << std::setprecision(static_cast<int>(precision)) << x;
             Parser::printLine();
         }
