+++ src/labs/inlab6/Parser.h
@@ -217,11 +217,11 @@ class Parser : public CommandLine<InLab6Inputs> {
             printLine();
             std::cout << "Coefficient Matrix (A):\n";
             printLine();
-            std::cout << std::setprecision(precision) << inputs.coefficients;
+            std::cout << std::setprecision(static_cast<int>(precision)) << inputs.coefficients;
             printLine();
             std::cout << "Constants Vector (b):\n";
             printLine();
-            std::cout << std::setprecision(precision) << inputs.constants;
+            std::cout << std::setprecision(static_cast<int>(precision)) << inputs.constants;
         }
     }
 };
