+++ src/labs/outlab3/outlab3.h
@@ -103,7 +103,7 @@ class OutLab3 : public Project<NewtonCotesInputs, Parser, Dictionary> {
             nlohmann::json json1;
             output1.toJSON(json1);
             outputsJSON[ruleKey(RULE_COMPOSITE_TRAPEZOIDAL)] = json1;
-            std::cout << "Integral over [a,b] using trapezoidal rule: " << std::setprecision(precision)
+            std::cout << "Integral over [a,b] using trapezoidal rule: " << std::setprecision(static_cast<int>(precision))
                       << output1.integral << std::endl;
             Parser::printLine();
         }
@@ -114,7 +114,7 @@ class OutLab3 : public Project<NewtonCotesInputs, Parser, Dictionary> {
             nlohmann::json json2;
             output2.toJSON(json2);
             outputsJSON[ruleKey(RULE_COMPOSITE_SIMPSONS)] = json2;
-            std::cout << "Integral over [a,b] using Simpson's rule:   " << std::setprecision(precision)
+            std::cout << "Integral over [a,b] using Simpson's rule:   " << std::setprecision(static_cast<int>(precision))
                       << output2.integral << std::endl;
             Parser::printLine();
         }
@@ -125,7 +125,7 @@ class OutLab3 : public Project<NewtonCotesInputs, Parser, Dictionary> {
             nlohmann::json json3;
             output3.toJSON(json3);
             outputsJSON[ruleKey(RULE_GAUSSIAN_QUADRATURE)] = json3;
-            std::cout << "Integral over [a,b] using Gauss-Legendre Quadrature:   " << std::setprecision(precision)
+            std::cout << "Integral over [a,b] using Gauss-Legendre Quadrature:   " << std::setprecision(static_cast<int>(precision))
                       << output3.integral << std::endl;
             Parser::printLine();
         }
